Fix for the `--single` option
Right now running list with `--single` option doesn't work correctly. I tries to load the file under a specified path, and when it doesn't find it, it crushes with an error:

```
Error: ENOENT: no such file or directory, open '/Users/dominik/Downloads/theory-463759986/whatever'
    at Error (native)
```

The above commit reverts the line that introduced the bug and fixes the behaviour for `--single` option.