Support for HTTPS in development added (#274)
* Support HTTPS

This should hopefully resolve #178. It uses the example given in the
`micro` examples of `with-https`. It has some rough edges still that
needed to be sanded down:

* Choosing a specific short flag name for ssl (and if the initial flag
should still be called ssl; I'm tempted to rename it to `https` or what
have you).

* The example always acks with 200 and the example object.

* It does not support compression and merely overwrites the previous
server var.

* Pass appropriate handler to https server

This also calls `compress' on the handler when the `unzipped' flag is
not present.

* Move https import to appropriate section

* Change SSL flag copy

* Merge in master to update package-lock.json

* Address conflicts

* Use master package-lock.json to avoid conflicts

* Make travis happy by updating ava snapshots

* Fix sha1 to sha512 flipping per npm 5.3.0 issue